+++ lib/league_stats.rb
@@ -1,0 +1,1 @@
+]
