+++ lib/team.rb
@@ -1,3 +1,5 @@
+#Team Stats
+
 class Team
 
   attr_reader :team_id, :franchise_id, :team_name, :abbreviation, :stadium, :link
